+++ R/dev_utils.R
@@ -1,3 +1,9 @@
 test.file <- function(...) {
   system.file("testdata", ..., package = "lightr")
 }
+
+int32_to_uint32 <- function(int32) {
+
+  ifelse(int32>=0, int32, 2^31-1 + int32)
+
+}