+++ R/parse_avantes_binary.R
@@ -143,3 +143,149 @@ lr_parse_abs <- lr_parse_trm
 #' @export
 #'
 lr_parse_roh <- lr_parse_trm
+
+#' @rdname lr_parse_trm
+#'
+#' @param specnum Integer representing the position of the spectrum to read in
+#'   the file. This option only makes sense for AvaSoft8 files and is ignored
+#'   in the other cases.
+#'
+#' @export
+#'
+#' @examples
+#' lr_parse_rfl8(system.file("testdata", "compare", "Avantes",
+#'                           "feather.RFL8",
+#'                           package = "lightr"),
+#'               specnum = 1)
+#' lr_parse_rfl8(system.file("testdata", "compare", "Avantes",
+#'                           "feather.RFL8",
+#'                           package = "lightr"),
+#'               specnum = 2)
+#'
+lr_parse_rfl8 <- function(filename, specnum) {
+
+  # File structure information provided courtesy of Avantes
+
+  f <- file(filename, "rb")
+  on.exit(close(f))
+
+  # HEADER
+  # always 'AVS82'
+  marker <- rawToChar(readBin(f, "raw", n = 5, endian = "little"))
+
+  # number of spectra in file
+  numspectra <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+  if (numspectra > 1 & missing(specnum)) {
+    specnum <- 1L
+    warning(
+      "This file contains multiple spectra and 'specnum' argument is missing. ",
+      "Returning the first spectrum by default.",
+      call. = FALSE
+    )
+  }
+  if (specnum > numspectra) {
+    stop("'specnum' is larger than the number of spectra in the input file",
+         call. = FALSE)
+  }
+
+  for (i in seq_len(numspectra)) {
+    # total length of the subfile
+    length <- int32_to_uint32(
+      readBin(f, "integer", size = 4, n = 1, endian = "little")
+    )
+
+    seqnum <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+    measmode <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+    bitness <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+    SDmarker <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+    # AvsIdentityType (75 bytes)
+    specID <- rawToChar(readBin(f, "raw", n = 10, endian = "little"))
+    userfriendlyname <- readBin(f, "raw", n = 64, endian = "little")
+    status <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+    # MeasConfigType (41 bytes)
+    m_startpixel <- readBin(f, "integer", size = 2, signed = FALSE, endian = "little")
+    m_stoppixel <- readBin(f, "integer", size = 2, signed = FALSE, endian = "little")
+    dark_inttime <- white_inttime <- scope_inttime <- readBin(f, "numeric", size = 4, endian = "little")
+    m_integrationdelay <- int32_to_uint32(
+      readBin(f, "integer", size = 4, endian = "little")
+    )
+    dark_average <- white_average <- scope_average <- int32_to_uint32(
+      readBin(f, "integer", size = 4, endian = "little")
+    )
+
+    ## DarkCorrectionType
+    m_enable <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+    m_forgetpercentage <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+    ## SmoothingType
+    dark_boxcar <- white_boxcar <- scope_boxcar <- readBin(f, "integer", size = 2, signed = FALSE, endian = "little")
+    m_smoothmodel <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+    m_saturationdetection <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+    ## TriggerType
+    m_mode <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+    m_source <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+    m_sourcetype <- readBin(f, "integer", size = 1, signed = FALSE, endian = "little")
+
+    ## ControlSettingsType
+    m_strobecontrol <- readBin(f, "integer", size = 2, signed = FALSE, endian = "little")
+    m_laserdelay <- int32_to_uint32(
+      readBin(f, "integer", size = 4, endian = "little")
+    )
+    m_laserwidth <- int32_to_uint32(
+      readBin(f, "integer", size = 4, endian = "little")
+    )
+    m_laserwavelength <- readBin(f, "numeric", size = 4, endian = "little")
+    m_storetoram <- readBin(f, "integer", size = 2, signed = FALSE, endian = "little")
+
+    timestamp <- readBin(f, "raw", n = 4, endian = "little")
+    SPCfiledate <- readBin(f, "raw", n = 4, endian = "little")
+
+    detectortemp <- readBin(f, "numeric", size = 4, endian = "little")
+    boardtemp <- readBin(f, "numeric", size = 4, endian = "little")
+
+    NTC2volt <- readBin(f, "numeric", size = 4, endian = "little")
+    ColorTemp <- readBin(f, "numeric", size = 4, endian = "little")
+    CalIntTime <- readBin(f, "numeric", size = 4, endian = "little")
+
+    fitdata <- readBin(f, "double", size = 8, n = 5, endian = "little")
+
+    comment <- intToUtf8(readBin(f, "raw", n = 130, endian = "little"))
+
+    len <- m_stoppixel - m_startpixel + 1
+
+    xcoord <- readBin(f, "numeric", 4, n = len, endian = "little")
+    scope <- readBin(f, "numeric", 4, n = len, endian = "little")
+    dark <- readBin(f, "numeric", 4, n = len, endian = "little")
+    reference <- readBin(f, "numeric", 4, n = len, endian = "little")
+
+    mergegroup <- intToUtf8(readBin(f, "raw", n = 10, endian = "little"))
+
+    data <- as.data.frame(cbind("wl" = xcoord,
+                                dark,
+                                "white" = reference,
+                                scope))
+    data$processed <- (data$scope - data$dark) / (data$white - data$dark) * 100
+
+    author <- NA
+    savetime <- NA # FIXME: extract this from SPCfiledate
+    specmodel <- NA
+
+    metadata <-  metadata <- c(author, savetime, specmodel, specID,
+                               dark_inttime, white_inttime, scope_inttime,
+                               dark_average, white_average, scope_average,
+                               dark_boxcar, white_boxcar, scope_boxcar)
+
+    if (specnum == i) {
+      return(list(data, metadata))
+    }
+
+  }
+}