+++ NEWS.md
@@ -2,8 +2,10 @@
 
 ## New features and major changes
 
-* `lightr` can correctly imports `TRM` files from Avasoft 6.0 (previously it 
-only supported files from Avasoft 7.0).
+* `lightr` can now import AvaSoft8 files (test file provided by M.D. Shawkey),
+via the function `lr_parse_rfl8()`.
+* `lightr` can correctly imports `TRM` files from AvaSoft 6.0 (previously it 
+only supported files from AvaSoft 7.0).
 * `lightr` can now import binary `.spc` files (via the `lr_parse_spc()` parser).
 This format is used by OceanInsight and CRAIC.